Tidy (JSONScript)
